Normalise type families in the type of an expression
Summary:
Before, the type of an expression, and the type of a variable
binding that expression used to be different in GHCi. The
reason being that types of bound variables were already normalised.
Now, both are normalised.

This implements the suggestions as given in Trac #10321
Also adds an expected output for test T10321

Reviewers: austin, simonpj, goldfire

Subscribers: thomie

Differential Revision: https://phabricator.haskell.org/D870

GHC Trac Issues: #10321