Run canonicalize pass after type-checker plugins are done
Wanted constraints are zonked before they are passed to the
type-checker plugins. Zonking makes all constraints
nonCanonical, which in turn leads to other passes of the
solver not picking up those constraints anymore; this lead to
situations such as #11525.

So after all type-checker plugins are run, we run one more
pass of the canonicalize stage.